@@ -1,16 +1,15 @@
 name: finance
 description: The fundamental package for financial computing with Dart.
-version: 0.1.0
-author: "Ismael Jiménez <ijimenezm@hotmail.es>"
+version: 0.1.1
 homepage: https://github.com/ismaelJimenez/finance
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.19.0 <4.0.0"
 
 dependencies:
   meta: ^1.1.7
 
 dev_dependencies:
-  test: '>=1.3.0 <1.6.8'
-  mockito: ^4.1.0
-  pedantic: ^1.8.0+1
+  flutter_lints: ^2.0.1
+  mockito: ^5.4.2
+  test: ^1.24.3
