@@ -43,10 +43,10 @@ class Finance {
   ///     http://www.oasis-open.org/committees/documents.php?wg_abbrev=office-formula
   ///     OpenDocument-formula-20090508.odt
   static num fv(
-      {@required num rate,
-      @required num nper,
-      @required num pmt,
-      @required num pv,
+      {required num rate,
+      required num nper,
+      required num pmt,
+      required num pv,
       bool end = true}) {
     final int when = end ? 0 : 1;
     final num temp = pow(1 + rate, nper);
@@ -98,9 +98,9 @@ class Finance {
   ///     http://www.oasis-open.org/committees/documents.php
   ///     ?wg_abbrev=office-formulaOpenDocument-formula-20090508.odt
   static num pmt(
-      {@required num rate,
-      @required num nper,
-      @required num pv,
+      {required num rate,
+      required num nper,
+      required num pv,
       num fv = 0,
       bool end = true}) {
     final int when = end ? 0 : 1;
@@ -135,9 +135,9 @@ class Finance {
   ///    but if ``rate = 0`` then::
   ///     fv + pv + pmt*nper = 0
   static num nper(
-      {@required num rate,
-      @required num pmt,
-      @required num pv,
+      {required num rate,
+      required num pmt,
+      required num pv,
       num fv = 0,
       bool end = true}) {
     final int when = end ? 0 : 1;
@@ -171,10 +171,10 @@ class Finance {
   /// of each period. (default=true).
   ///
   static num ipmt(
-      {@required num rate,
-      @required num per,
-      @required num nper,
-      @required num pv,
+      {required num rate,
+      required num per,
+      required num nper,
+      required num pv,
       num fv = 0,
       bool end = true}) {
     final num totalPmt = pmt(rate: rate, nper: nper, pv: pv, fv: fv, end: end);
@@ -189,10 +189,10 @@ class Finance {
   //  function to not interfere with the 'fv' keyword argument within the 'ipmt'
   //  function.  It is the 'remaining balance on loan'.
   static num _rbl(
-      {@required num rate,
-      @required num per,
-      @required num pmt,
-      @required num pv,
+      {required num rate,
+      required num per,
+      required num pmt,
+      required num pv,
       bool end = true}) {
     return fv(rate: rate, nper: per - 1, pmt: pmt, pv: pv, end: end);
   }
@@ -216,10 +216,10 @@ class Finance {
   /// of each period. (default=true).
   ///
   static num ppmt(
-      {@required num rate,
-      @required num per,
-      @required num nper,
-      @required num pv,
+      {required num rate,
+      required num per,
+      required num nper,
+      required num pv,
       num fv = 0,
       bool end = true}) {
     final num total = pmt(rate: rate, nper: nper, pv: pv, fv: fv, end: end);
@@ -263,10 +263,10 @@ class Finance {
   ///    http://www.oasis-open.org/committees/documents.php?wg_abbrev=office-formula
   ///    OpenDocument-formula-20090508.odt
   static num pv(
-      {@required num rate,
-      @required num nper,
-      @required num pmt,
-      @required num fv,
+      {required num rate,
+      required num nper,
+      required num pmt,
+      required num fv,
       bool end = true}) {
     final int when = end ? 0 : 1;
     final num temp = pow(1 + rate, nper);
@@ -336,10 +336,10 @@ class Finance {
   ///    http://www.oasis-open.org/committees/documents.php?wg_abbrev=office-formula
   ///    OpenDocument-formula-20090508.odt
   static num rate(
-      {@required num nper,
-      @required num pmt,
-      @required num pv,
-      @required num fv,
+      {required num nper,
+      required num pmt,
+      required num pv,
+      required num fv,
       bool end = true,
       num guess = 0.1,
       num tol = 1e-6,
@@ -380,13 +380,13 @@ class Finance {
   /// ----------
   ///    .. [G] L. J. Gitman, "Principles of Managerial Finance, Brief," 3rd ed.,
   ///       Addison-Wesley, 2003, pg. 346.
-  static num npv({@required num rate, @required List<num> values}) {
+  static num npv({required num rate, required List<num> values}) {
     return List<int>.generate(values.length, (int index) => index)
         .map((int index) => values[index] / pow(1 + rate, index))
         .fold(0, (num p, num c) => p + c);
   }
 
-  static num _npvPrime({@required num rate, @required List<num> values}) {
+  static num _npvPrime({required num rate, required List<num> values}) {
     return List<int>.generate(values.length, (int index) => index)
         .map((int index) => -index * values[index] / pow(1 + rate, index + 1))
         .fold(0, (num p, num c) => p + c);
@@ -427,7 +427,7 @@ class Finance {
   ///    g'(r) is the derivative with respect to r.
 
   static num irr(
-      {@required List<num> values,
+      {required List<num> values,
       num guess = 0.1,
       num tol = 1e-6,
       num maxIter = 100}) {
